1. Some bugfixes for StswToggleSwitch. 2. StswTreeView no longer have UsesSelectionItems property since it have SelectedItem property and cannot select multiple items. 3. Some changes to StswAdaptiveBox (auto value is now Type equal to null + Type changes itself from null to different now). 4. StswFilter uses StswAdaptiveBox now. 5. StswNumericBox uses a decimal typed Value instead of double typed Value - it should fix problem with precision. 6. Input controls and StswProgressBar are now stretched by default instead of topped. 7. StswLogPanel do not tab items by default. 8. Enum StswFilterType has been removed since StswAdaptiveType has its role now.

+++ StswExpress/Controls/Input/StswTextEditor.xaml.cs
@@ -159,7 +159,7 @@ public class StswTextEditor : RichTextBox
         if (fontSize != null)
         {
             if (temp != DependencyProperty.UnsetValue)
-                fontSize.Value = Convert.ToDouble(temp);
+                fontSize.Value = Convert.ToDecimal(temp);
             else
                 fontSize.Value = null;
         }
